# openapi.yaml
openapi: 3.1.0
info:
  title: Enterprise E-Commerce API
  version: 0.1.0
  description: REST API for catalog, checkout, orders, and payments.
servers:
  - url: http://localhost:8000
    description: Local development
paths:
...
  /api/v1/products:
    get:
      summary: List products
      operationId: listProducts
      tags: [catalog]
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: category
          in: query
          description: Filter by primary category slug
          schema:
            type: string
        - name: in_stock
          in: query
          description: When true, only in-stock products
          schema:
            type: boolean
        - name: on_sale
          in: query
          description: When true, only discounted products
          schema:
            type: boolean
        - name: size
          in: query
          description: Filter by variant size (repeatable)
          schema:
            type: array
            items:
              type: string
          style: form
          explode: true
        - name: color
          in: query
          description: Filter by variant color or camouflage
          schema:
            type: array
            items:
              type: string
          style: form
          explode: true
        - name: price_min
          in: query
          description: Minimum price in cents
          schema:
            type: integer
            minimum: 0
        - name: price_max
          in: query
          description: Maximum price in cents
          schema:
            type: integer
            minimum: 0
        - name: sort
          in: query
          description: Sort order (popular = best sellers by quantity in last 90 days)
          schema:
            type: string
            enum: [default, popular, price_asc, price_desc, name_asc, name_desc]
            default: default
      responses:
        "200":
          description: Paginated product list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductListResponse"
...
  /api/v1/products/{slug}:
    get:
      summary: Get product by slug
      operationId: getProduct
      tags: [catalog]
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Product detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Product not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/v1/admin/catalog/products:
    get:
      summary: Admin list products (all statuses)
      operationId: adminListProducts
      tags: [admin]
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: status
          in: query
          schema:
            type: string
            enum: [draft, active, archived]
        - name: q
          in: query
          schema:
            type: string
            maxLength: 200
          description: Search by product name, slug, or variant SKU (all statuses)
        - name: category_id
          in: query
          schema:
            type: string
            format: uuid
          description: Filter products by primary category
        - name: uncategorized
          in: query
          schema:
            type: boolean
            default: false
          description: When true, list only products without a category
      responses:
        "200":
          description: Paginated admin product list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminProductListResponse"
    post:
      summary: Admin create product
      operationId: adminCreateProduct
      tags: [admin]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminCreateProductRequest"
      responses:
        "201":
          description: Product created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminProductSchema"
  /api/v1/admin/catalog/categories:
    get:
      summary: Admin list categories (all)
      operationId: adminListCategories
      tags: [admin]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Category list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminCategoryListResponse"
    post:
      summary: Admin create category
      operationId: adminCreateCategory
      tags: [admin]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminCreateCategoryRequest"
      responses:
        "201":
          description: Category created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ProductVariant:
      type: object
      required: [id, sku, name, price_cents, in_stock, is_default, sort_order, attributes]
      properties:
        id:
          type: string
          format: uuid
        sku:
          type: string
        name:
          type: string
        price_cents:
          type: integer
        wholesale_price_cents:
          type: integer
          description: Omitted for non-wholesaler API consumers
        in_stock:
          type: boolean
        is_default:
          type: boolean
        sort_order:
          type: integer
        attributes:
          type: object
          additionalProperties:
            type: string
    Product:
      type: object
      required: [id, name, slug, price_cents, currency, in_stock]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        price_cents:
          type: integer
        compare_at_price_cents:
          type: integer
          nullable: true
        currency:
          type: string
          example: RUB
        in_stock:
          type: boolean
        category_id:
          type: string
          format: uuid
          nullable: true
        description:
          type: string
          nullable: true
        image_url:
          type: string
          nullable: true
        variants:
          type: array
          items:
            $ref: "#/components/schemas/ProductVariant"
    Category:
      type: object
      required: [id, slug, name, is_active, sort_order]
      properties:
        id:
          type: string
          format: uuid
        slug:
          type: string
        name:
          type: string
        description:
          type: string
          nullable: true
        parent_id:
          type: string
          format: uuid
          nullable: true
        is_active:
          type: boolean
        sort_order:
          type: integer
        product_count:
          type: integer
          default: 0
...
    ProductListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
...
    ErrorResponse:
      type: object
      properties:
        detail:
          type: string
...
    AdminProductSchema:
      type: object
      required: [id, name, slug, price_cents, currency, in_stock, status]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        price_cents:
          type: integer
        compare_at_price_cents:
          type: integer
          nullable: true
        currency:
          type: string
        in_stock:
          type: boolean
        status:
          type: string
          enum: [draft, active, archived]
        category_id:
          type: string
          format: uuid
          nullable: true
        description:
          type: string
          nullable: true
        image_url:
          type: string
          nullable: true
        variants:
          type: array
          items:
            $ref: "#/components/schemas/ProductVariant"
    AdminProductListResponse:
      type: object
      required: [items, total, page, limit]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/AdminProductSchema"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    AdminCreateProductRequest:
      type: object
      required: [name, slug, sku, price_cents]
      properties:
        name:
          type: string
        slug:
          type: string
        sku:
          type: string
        price_cents:
          type: integer
        currency:
          type: string
          default: USD
        status:
          type: string
          enum: [draft, active, archived]
          default: draft
        compare_at_price_cents:
          type: integer
          nullable: true
        category_id:
          type: string
          format: uuid
          nullable: true
    AdminCategoryListResponse:
      type: object
      required: [items, total]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Category"
        total:
          type: integer
    AdminCreateCategoryRequest:
      type: object
      required: [slug, name]
      properties:
        slug:
          type: string
        name:
          type: string
        description:
          type: string
          nullable: true
        parent_id:
          type: string
          format: uuid
          nullable: true
        is_active:
          type: boolean
          default: true
        sort_order:
          type: integer
          default: 0
...